+++ karma.conf.template.js
@@ -1,10 +1,13 @@
 'use strict'
+/* eslint-disable @typescript-eslint/no-var-requires */
+
+const path = require('path')
 
 module.exports = (config) => ({
   // Increase timeout in case connection in CI is slow
   captureTimeout: 120000,
-  browserNoActivityTimeout: 30000,
-  browserDisconnectTimeout: 30000,
+  browserNoActivityTimeout: 300000,
+  browserDisconnectTimeout: 300000,
   browserDisconnectTolerance: 3,
 
   // frameworks to use
@@ -19,7 +22,27 @@ module.exports = (config) => ({
   // test results reporter to use
   // possible values: 'dots', 'progress'
   // available reporters: https://npmjs.org/browse/keyword/karma-reporter
-  reporters: ['progress'],
+  reporters: ['progress', 'junit', 'coverage-istanbul'],
+
+  coverageIstanbulReporter: {
+    // reports can be any that are listed here: https://github.com/istanbuljs/istanbuljs/tree/73c25ce79f91010d1ff073aa6ff3fd01114f90db/packages/istanbul-reports/lib
+    reports: ['html', 'text', 'cobertura', 'text-summary'],
+    // base output directory. If you include %browser% in the path it will be replaced with the karma browser name
+    dir: path.join(__dirname, 'coverage', '%browser%'),
+    // Combines coverage information from multiple browsers into one report rather than outputting a report
+    // for each browser.
+    combineBrowserReports: false,
+    // if using webpack and pre-loaders, work around webpack breaking the source path
+    fixWebpackSourcePaths: true,
+    // Omit files with no statements, no functions and no branches covered from the report
+    skipFilesWithNoCoverage: true,
+    verbose: true // output config used by istanbul for debugging
+  },
+
+  junitReporter: {
+    outputDir: 'test-results', // results will be saved as $outputDir/$browserName.xml
+    useBrowserName: true // add browser name to report and classes names
+  },
 
   // Continuous Integration mode
   // if true, Karma captures browsers, runs the tests and exits
@@ -36,7 +59,7 @@ module.exports = (config) => ({
   logLevel: config.LOG_INFO,
 
   // enable / disable watching file and executing tests whenever any file changes
-  autoWatch: true,
+  autoWatch: false,
 
   // Concurrency level
   // how many browser should be started simultaneous